+++ resources/release-notes.md
@@ -1,5 +1,6 @@
 # App
 
 - Add placeholders for CPT settings with dynamic defaults
+- Move `cropResults` definition from `mag.app.control.Control` to `mag.app.internal.cropResults`
 - Disable "Save Figures" and "Close Figures" buttons after using "Close Figures"
 - Fix issues with validation of CPT primary and secondary modes, and range patterns (`mag.app.control.CPT`)